requirements.txt
streamlit==1.29.0
groq>=0.25.0
sentence-transformers>=4.1.0
chromadb==0.4.18
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
spacy==3.7.2
nltk==3.8.1
numpy==1.24.3
pandas==2.1.4
typing-extensions>=4.13.0
scikit-learn==1.3.2